CrearFicheroyDirectorio terminado

+++ src/com/ipartek/formacion/javalibro/ficheros/CrearFicheroYDirectorio.java
@@ -30,42 +30,49 @@ public class CrearFicheroYDirectorio {
 	static FileReader fr = null;
 	static BufferedReader br = null;
 	static String linea = "";
-	static Scanner sc= new Scanner(System.in);
-	static File rutaFichero=null;
+	static Scanner sc = new Scanner(System.in);
+	static File rutaFichero = null;
 
 	public static void crearFichero() {
 
 		System.out.println("Escribe la ruta de la carpeta con formato: 'C:\\\\yyyyyy'");
-		carpeta= new File(sc.nextLine());
-		
-		if(carpeta.exists()) {
-			
+		carpeta = new File(sc.nextLine());
+
+		if (carpeta.exists()) {
+
 			System.out.println("Ya existe una carpeta con ese nombre y en esa ruta");
-			
-		}else {
-			
+
+		} else {
+
 			carpeta.mkdir();
-			
+
 		}
-		
+
 	}
 
 	public static void escribirFichero() {
 
 		try {
+			System.out.println();
 			System.out.println("Escribe el nombre del txt(sin escribir ruta ni .txt)");
-			
-			rutaFichero= new File(carpeta+"\\"+sc.nextLine()+".txt");
-			
-			fw = new FileWriter(rutaFichero);
-			bw = new BufferedWriter(fw);// Siempre vamos a trabajar con el Buffer
-			
-			System.out.println("Escribe dentro del txt");
-			bw.write(sc.nextLine());
 
-			System.out.println("Fichero creado con exito");
-			System.out.println("------------------------");
+			rutaFichero = new File(carpeta + "\\" + sc.nextLine() + ".txt");
 
+			if (rutaFichero.exists()) {
+
+				System.out.println("Ya existe un fichero con ese nombre");
+			} else {
+				fw = new FileWriter(rutaFichero);
+				bw = new BufferedWriter(fw);// Siempre vamos a trabajar con el Buffer
+				
+				System.out.println();
+				System.out.println("Escribe dentro del txt");
+				bw.write(sc.nextLine());
+
+				System.out.println("Fichero creado con exito");
+				System.out.println("------------------------");
+
+			}
 		} catch (IOException e) {
 
 			System.out.println("No se puede escribir en el fichero");
@@ -81,6 +88,7 @@ public class CrearFicheroYDirectorio {
 				e.printStackTrace();
 			}
 		}
+
 	}
 
 	public static void main(String[] args) {
